refactor(types): on stable ground now
