--- src/stores/selected-song.store.ts
@@ -1,6 +1,0 @@
-import { writable } from 'svelte/store';
-import { SongsterrPartialMetadata } from '$lib/types';
-
-export const selectedSongToDownload = writable<
-  SongsterrPartialMetadata | undefined
->();
