+++ src/ui/ByLink/Form.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { enhance } from '$app/forms';
-  import { selectedSongToDownload } from '../../stores/selected-song.store';
+  import { selectedSongToDownload } from '../../lib/stores/selected-song.store';
   import {
     setValidationMessage,
     clearValidationMessage
@@ -12,7 +12,7 @@
   import { sample } from 'lodash-es';
   import { placeholderSongUrls } from '$lib/constants/placeholder-songsterr-url';
   import { toastError } from '$lib/utils/toast.util';
-  import { SongsterrPartialMetadata } from '$lib/types';
+  import type { SongsterrPartialMetadata } from '$lib/types';
 
   let selectedSong: SongsterrPartialMetadata | undefined;
