+++ src/ui/BySearch/SearchResult.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { selectedSongToDownload } from '../../stores/selected-song.store';
-  import { SongsterrPartialMetadata } from '$lib/types';
+  import { selectedSongToDownload } from '$lib/stores/selected-song.store';
+  import type { SongsterrPartialMetadata } from '$lib/types';
 
   export let searchResult: SongsterrPartialMetadata;
   export let index: number = 0;
