+++ src/ui/common/TabMenu.svelte
@@ -9,12 +9,12 @@
   import {
     activeTabMenuIndex,
     type ITabMenuIndex
-  } from '../../stores/tab-menu.store';
+  } from '$lib/stores/tab-menu.store';
 
   import { createClassNames } from '$lib/utils/css';
   import ByLink from '../ByLink/Form.svelte';
   import BySearch from '../BySearch/Form.svelte';
-  import { selectedSongToDownload } from '../../stores/selected-song.store';
+  import { selectedSongToDownload } from '$lib/stores/selected-song.store';
 
   function setActiveTabIndex(idx: ITabMenuIndex): void {
     activeTabMenuIndex.set(idx);
